moved model to another folder, changed design of project and sort pipe
