--- src/app/pipes/filter-category.pipe.ts
@@ -1,12 +1,0 @@
-import { Pipe, PipeTransform } from '@angular/core';
-
-@Pipe({
-  name: 'filterCategory'
-})
-export class FilterCategoryPipe implements PipeTransform {
-
-  transform(value: any, args?: any): any {
-    return null;
-  }
-
-}
