+++ src/app/projects/project-comments/project-comments.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { ProjectServiceService } from '../../core/project-service/project-service.service';
-import { Project } from '../models/project.model';
 
 @Component({
   selector: 'comments',
