+++ src/app/projects/project-edit/project-edit.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { ProjectServiceService } from '../../core/project-service/project-service.service';
 import { Router } from '@angular/router';
-import { Project } from '../models/project.model';
+import { Project } from '../../models/project.model';
 import { AuthService } from '../../core/auth-service/auth-service.service';
 import { NgForm } from '@angular/forms';
 
@@ -65,7 +65,7 @@ export class ProjectEditComponent implements OnInit {
           this.errorMessage = error;
         });
     } else {
-      this.putProject.patchProject(this.tempId.id, projectEdit)
+      this.putProject.putProject(this.tempId.id, projectEdit)
         .subscribe(
         (response) => {
           this.router.navigate(['/projects/' + this.tempId.id]);
