+++ src/app/projects/project-rating/project-rating.component.css
@@ -1,10 +1,10 @@
 .rating_block {
-  text-align: left;
   font-family: sans-serif;
   color: #ccc;
   padding: 0 0 0 5px;
 }
 
 .rating_info {
+    vertical-align: top;
   font-size: 16px;
 }
