+++ src/app/projects/projects-list/projects-list.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, Output, EventEmitter } from '@angular/core';
 import { ProjectServiceService } from '../../core/project-service/project-service.service';
+import { AuthService } from '../../core/auth-service/auth-service.service';
 
 @Component({
   selector: 'app-projects-list',
@@ -8,7 +9,8 @@ import { ProjectServiceService } from '../../core/project-service/project-servic
 })
 export class ProjectsListComponent implements OnInit {
 
-  constructor(private projectsData: ProjectServiceService) { }
+  constructor(private projectsData: ProjectServiceService,
+              private authService: AuthService) { }
 
   projects;
   userProjects = false;
@@ -38,6 +40,7 @@ export class ProjectsListComponent implements OnInit {
   }
 
   ngOnInit() {
+    console.log(this.authService.getRole());
     this.projects = this.projectsData.getProjects();
   }
